requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
pydantic<2.0